Update eager secret logic (#669)
## Overview
Currently Admin will add an eager secret to all eager tasks.  It detects that a task is eager with this bit.  In [this change](https://github.com/flyteorg/flytekit/pull/3148) though, we've added a cleanup task as a failure node to the eager workflow.  Please see the other PR for more information but this cleanup task is responsible for checking that all the executions launched by an eager task have terminated.  It will abort executions in non-terminal states.  But to do so, it also will need access to the eager secret.  This PR adds that request.

## Test Plan
Deployed and tested on staging.

## Rollout Plan (if applicable)
Test again on canary and then roll out to prod.

## Upstream Changes
Should not be upstreamed.

- [ ] To be upstreamed to OSS

## Issue
*TODO: Link Linear issue(s) using [magic words](https://linear.app/docs/github#magic-words). `fixes` will move to merged status, while `ref` will only link the PR.*

## Checklist
* [x] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation