update failure reason (#682)
## Overview
"terminated in the background, manually" is a confusing error message for users when their pods could be terminated by the cloud provider during things such as node upgrades

ref: https://unionai.slack.com/archives/C08FRASBR4G/p1741024712560879

Showing users how many retries are left could be a follow-up. If this is something users want, it could be worth a quick check in w/ design and FE team on what this should look like + how much time it'd take. This should be a small eventing change on the BE to include "how many retries (system and/or user) are left".

## Test Plan
n/a

## Rollout Plan (if applicable)
roll out to all tenants

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [x] To be upstreamed to OSS

## Issue
*TODO: Link Linear issue(s) using [magic words](https://linear.app/docs/github#magic-words). `fixes` will move to merged status, while `ref` will only link the PR.*

## Checklist
* [ ] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation