update named_entity_metadata w/ has_trigger for lps + update named_entity_metadata on lp create and update (#687)
## Overview
When toggling "Has Trigger" for launch plans in console, the frontend has a cached query of named_entities and then makes queries for each launch plan to determine if has trigger.

This change adds a "has_trigger" field to named_entity_metadata such that console won't have to query for each launch plan. On all launch plan creates, a named_entity_metadata entry will be updated if the launch plan has a trigger. This means that named_entity_metadata has_trigger will match that of the most recently registered launch plan.

## Test Plan
- added unit tests
- ran in dogfood: https://dogfood.cloud-staging.union.ai/api/v1/named_entities/3/flytesnacks/development?limit=1000
- confirmed migration set has_trigger for appropriate named_entities + newly registered launch plans 

## Rollout Plan (if applicable)
- roll out to all tenants. 

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [ ] To be upstreamed to OSS

## Issue
fixes: https://linear.app/unionai/issue/BB-2781/allow-filtering-named-entity-list-requests-by-has-trigger

## Checklist
* [x] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation