Add an empty config in the connector-configmap (#694)
## Overview
https://unionai.slack.com/archives/C06RE4ZR5QQ/p1742487674524919
`connector-config` config map is out of sync in the argo pipeline because the union-operator will update the config map once it's created.

Note: This PR is a workaround. It will reset the ConfigMap to empty, but the `union-operator` will immediately update the config once it restarts

## Test Plan
https://buildkite.com/unionai/managed-cluster-staging-sync/builds/7874#
1. Remove the config map and run the pipeline. connector-config will be created.
2. Update the config map and rerun the pipeline again

## Rollout Plan (if applicable)
All tenants

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [ ] To be upstreamed to OSS

## Issue
https://linear.app/unionai/issue/BB-4943/create-the-connector-config-if-it-does-not-exist

## Checklist
* [ ] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation