Making ArrayNode taskPhaseVersion increment retries configurable (#698)
## Overview
This PR makes the ArrayNode `taskPhaseVersion` increment retries configurable. This value is used to resend `TaskExecutionEvents` in the case of ArrayNode eventing failures. It is necessary because ArrayNode uses the `ExternalResourceInfo` fields to populate subNode execution metadata, all of these are included in a singular event. So the `TaskPhaseVersion` is not necessarily correlated to a single phase (as subNode phases can change). To ensure these are populated we need to increment.

## Test Plan
The default value is the current value (ie. 3) so there should be no concerns about merging.

## Rollout Plan (if applicable)
This will be rolled out immediately using the `managed-cluster-*` pipelines.

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [x] To be upstreamed to OSS

## Issue
https://linear.app/unionai/issue/BB-4957/configurable-arraynode-taskphaseversion-attempt-increments

## Checklist
* [ ] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation