Add yaml tags to LogConfig fields (#705)
## Overview
Add `omitempty` to the yaml tag for LogConfig, so that the app controller won't create an empty log config in the `connector-config` 
https://github.com/unionai/cloud/blob/65673b761170a798ff8f1099f590b4e2e181a3ed/operator/apps/controller.go#L241

## Test Plan
<img width="1677" alt="Screenshot 2025-04-07 at 9 00 43 PM" src="https://github.com/user-attachments/assets/e71ab174-6835-4220-a3e4-2e0211965447" />
https://dogfood-gcp.cloud-staging.union.ai/console/projects/flytesnacks/domains/development/executions/ad46wpjcjx8bk55sh48z/nodeId/n0/nodes

## Rollout Plan (if applicable)
all tenants

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [ ] To be upstreamed to OSS

## Issue
NA

## Checklist
* [ ] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation