Managed Images Pod Mutator (#729)
## Overview
This PR adds a pod mutator that is configured with certain managed images and can configure the pod to use the right resolved image for that region and cluster configuration (does it have nydus enabled or not... etc.)

## Test Plan
Deployed to serving-mvp
Created a deployment with image `managed.cr.union.ai/sglang:stable` and see it got replaced with the public ECR one and tolerations + node selector added to the pod

## Rollout Plan (if applicable)
Need to update cloud repo and rollout managed-cluster sync

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [ ] To be upstreamed to OSS

## Issue
fixes https://linear.app/unionai/issue/AI-395/inject-nydus-tolerationsnode-selector

## Checklist
* [ ] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation