Increase max length for node id (#732)
## Overview
Increase to 30 to accommodate action names

## Test Plan
- [x] Run in v2 devbox
- [x] Run [functional tests](https://buildkite.com/unionai/functional-tests-staging-run/builds/490#01974137-e143-4196-bf55-44c2a20aacd8)

## Rollout Plan (if applicable)
Merge and bring to unionai/cloud

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [ ] To be upstreamed to OSS

## Issue
ref BB-5370

## Checklist
* [ ] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation