include metadata in cache put failure statuses (#736)
## Overview
task 1 and task 2 have the same cache key and no cache hits prior to those runs
task 2 starts before task 1 finishes and writes to cache
task 2 executes but then fails writing to cache as there's already an existing entry

issue is admin was expecting catalog key to be set in taskMetadata to update the closure to include the cache status. This wasn't getting set on the status on put failures, so it'd always be the default cache status disabled.
https://github.com/unionai/flyte/blob/b6482757bd316476cae3113712a9284ce3fae9dc/flyteadmin/pkg/repositories/transformers/node_execution.go#L258

## Test Plan
updated unit tests
confirmed in sandbox that cache put failure shows up instead of cache disabled

![image](https://github.com/user-attachments/assets/84f45ed3-7581-43c1-ab06-7b8356532176)


## Rollout Plan (if applicable)
fine to roll out to all tenants

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [x] To be upstreamed to OSS

## Issue
fixes: https://linear.app/unionai/issue/BB-5623/cachine-race-condition-not-recognized-in-ui

## Checklist
* [x] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation