docs: Add clarification on required Kafka version field in cloud_events configuration for FlyteAdmin (#6502)
The current documentation for the FlyteAdmin cloudEvents Kafka configuration does not mention that the `version` field under the `kafka` block is required. 

If this field is omitted, attempting to start FlyteAdmin with Kafka cloud events enabled leads to a panic error with the message "caught panic: invalid version".

An example of a minimal valid configuration is as follows:
```
cloudEvents:
  enable: true
  kafka:
    brokers: 127.0.0.1:9092
    version: my-kafka-version
  eventsPublisher:
    eventTypes:
      - all
    topicName: myTopic
  type: kafka
```
This change aims to help users avoid runtime errors and improve the ease of adoption for this feature.

Signed-off-by: Daaa <108403536+daadc@users.noreply.github.com>