Inline refresh cookies on REST requests (#762)
## Overview
Currently when idtoken/accesstoken (stored in cookies) expire, we redirect users to /login. The issue is that this can happen while the user is navigating around which isn't a good experience. Ideally we just automatically refresh cookies if the refresh token is still valid to do so.

## Test Plan
Deploy `org-staging-sync` and leave a page open, we shouldn't see it redirect to `/login` for 12 hours

## Rollout Plan (if applicable)
Revert and run `org-staging-sync`

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [X] To be upstreamed to OSS

## Issue
https://linear.app/unionai/issue/BB-5757/cookies-should-be-refreshed-inline

## Checklist
* [ ] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation