Clean up default fetcher for embedded secret manager (#778)
## Overview
We added a default fetcher for embedded secret manager to unblock Nimble. However, it turned out we can simply use OSS secret manager. This PR cleans up the default fetcher since it's not only not in use but also incomplete.

## Test Plan
Tested in staging.

## Rollout Plan (if applicable)
Run managed-cluster pipeline

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [ ] To be upstreamed to OSS

## Issue
N/A

## Checklist
* [ ] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation