consider orphaned last accessed at to prevent scale down deleting env inf loop (#790)
## Overview
we can enter a scenario in which a scale down is triggered on an env before the first replica sends a heartbeat to update its status to healthy. Since we previously weren't considering the last accessed timestamp for orphaned pods, the env would get deleted (deleting the orphaned, but probably healthy, worker). This would then cause for a inf loop of the env getting deleted and spun up.

## Test Plan
updated unit tests

## Rollout Plan (if applicable)
managed-all

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [] To be upstreamed to OSS

## Issue
ref: https://linear.app/unionai/issue/BB-5849/actors-autoscaling

## Checklist
* [ ] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation