Concurrency bridge attempt 2.2 (#791)
## Overview
This is an update to v2 of the actor bridge. v1 remains unchanged.  The patterns are roughly the same - just simplified a bit in our minds.  Updates to documentation and design docs to come.  Also there is a litany of additional work that will be needed following this PR.

## Test Plan
Has been tested on demo.

## Rollout Plan (if applicable)
None.

## Upstream Changes
This should not be upstreamed.  There is some discussion about open sourcing the executor side of actors though.
- [ ] To be upstreamed to OSS

## Issue
https://linear.app/unionai/issue/PRD-249/get-async-eager-tasks-running-concurrently

## Checklist
* [ ] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation