Remove the redundant suffix in the dashboard name (#786)
## Overview
To remove the (User logs) from the display name of the dashboard link

<img width="1126" height="369" alt="Screenshot 2025-08-15 at 2 06 01 PM" src="https://github.com/user-attachments/assets/1270d041-16f5-4593-a3d0-7e615f8a4f7a" />

## Test Plan
unit test

## Rollout Plan (if applicable)
All tenants

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [ ] To be upstreamed to OSS

## Issue
NA

## Checklist
* [ ] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation