improve orphan worker scale down (#804)
## Overview
Before this change, we defined workers as just orphaned or healthy. When workers are created they are set to orphaned and their state is updated to healthy upon a heartbeat connection - sets back to orphaned when this connection is disrupted.

We scale down expired orphaned workers after a grace period (30 seconds to allow for workers to reconnect so any running tasks can be salvaged) below the min replica count to better ensure we have the min replica count of healthy workers. 

However this presents an issue of worker pods that take a long time to spin up as they're garbage collected prior to spinning up, potentially creating a long or permanent loop of terminations and spin ups. Tasks would stay in resources not available so would just be waiting on any configured timeout. 

Ideally we'd want to quickly terminate unhealthy, disconnected workers with a short orphan TTL but also enable worker pods ample time to spin up. We can do this by setting the initial worker state to initializing and set a long initializing ttl.

## Test Plan
added unit tests
ran locally

## Rollout Plan (if applicable)
managed-all

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [] To be upstreamed to OSS

## Issue
fixes: https://linear.app/unionai/issue/BB-5910/clean-up-scale-down-w-actor-orphaned

## Checklist
* [ ] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation