fix(k8s): Update task phase version while task is running (#817)
## Overview
- Fix phase version update condition in MaybeUpdatePhaseVersion to properly handle all phases, not just non-running phases
- Add phase check before processing IDE logs to ensure vscode readiness is only evaluated when pod is running or later phases

## Test Plan
Run this locally, check if the button shows up 

## Rollout Plan (if applicable)
All tenants

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [ ] To be upstreamed to OSS

## Issue
NA

## Checklist
* [ ] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation