Add log context for the pytorch task (#821)
## Overview
Add log context for the pytorch task

## Test Plan
Run a pytorch task

<img width="1681" height="1182" alt="Screenshot 2025-09-19 at 4 53 20 PM" src="https://github.com/user-attachments/assets/92a7b4ad-60e5-4d45-8e29-4aed4944549a" />

## Rollout Plan (if applicable)
managed-cluster-*-sync

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [ ] To be upstreamed to OSS

## Issue
NA

## Checklist
* [ ] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation