set node phase timed out on execution time outs when out of retries (#831)
## Overview
Set phase info to TimedOut if execution timeout and out of retries. Fixes left side of UI in v2 to show TimedOut instead of Failed.

## Test Plan
- updated unit test
- ran locally

https://dogfood-gcp.cloud-staging.union.ai/v2/runs/project/flytesnacks/domain/development/rv6k7zr8tcp2qkd8xttc?i=6ek0le7kkfx5delj2p01wfx7j&actionSearch= (along w/ https://github.com/unionai/cloud/pull/13053)

## Rollout Plan (if applicable)
managed-all

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [x] To be upstreamed to OSS

## Issue
ref: https://linear.app/unionai/issue/BB-6046/clean-up-time-outs

## Checklist
* [ ] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation