fix: Support string values for file secret mounts (#834)
## Overview
Support string values for file secret mounts

## Test Plan
Created a secret from a string value and then ran a task with this secret.

## Rollout Plan (if applicable)
managed-cluster-*-sync

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [ ] To be upstreamed to OSS

## Issue
https://unionai.slack.com/archives/C06RE4ZR5QQ/p1759954939368169?thread_ts=1759867415.896959&cid=C06RE4ZR5QQ

## Checklist
* [ ] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation