Add IsV2Task check for port config (#840)
## Overview
objectMeta.Labels is always empty when creating the Kubernetes Pod object. Instead, check the environment variable to determine whether the task is V2, and modify the Pod for readiness probe accordingly

## Test Plan
Run a v2 task with debugger

## Rollout Plan (if applicable)
managed-cluster-*-sync

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [ ] To be upstreamed to OSS

## Issue
https://unionai.slack.com/archives/C07QTRXMH8C/p1760473332437389

## Checklist
* [ ] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation