fix(k8s_resolver): Add error handling for watcher events (#839)
## Overview
Handle different event types accordingly based on the event type in the Kubernetes resolver

## Test Plan
Deploy a new k8s resolver to staging and check the log

## Rollout Plan (if applicable)
managed-cluster-*-sync

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [ ] To be upstreamed to OSS

## Issue
https://unionai.slack.com/archives/C03SJ1VH8MV/p1760442455038109

## Checklist
* [ ] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation