correctly cast timeout for grace period seconds (#850)
## Overview
some v2 pods in demo were stuck in terminating. This was due to the deletionTimeStamp being in the far future due to terminationGracePeriodSeconds being set to 3600000000000

## Test Plan
ran locally

## Rollout Plan (if applicable)
managed-all 

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [x] To be upstreamed to OSS

## Issue
fixes: https://linear.app/unionai/issue/BB-6136/demo-pods-stuck-terminating

## Checklist
* [ ] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation