Initializing TaskContext on task assignment to actor replica (#851)
## Overview
We were not initializing the `TaskContext` when the task was assigned to an actor replica. This resulted in requiring multiple "heartbeats" to detect when the tasks first phase was terminal. For example, the first heartbeat on terminal phase resulted in creating the `TaskContext` and the second populated it. In this PR, we create the `TaskContext` immediately, so the first terminal phase is correctly processed. I strongly suspect somewhere this initialization got removed.

## Test Plan
Tested locally, but will test on cloud before merging.

## Rollout Plan (if applicable)
`managed-cluster-*-sync-all`

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [ ] To be upstreamed to OSS

## Issue
https://linear.app/unionai/issue/BB-6178/v2-reuseable-containers-can-have-slow-execution-times

## Checklist
* [ ] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation