Using current time on phase transition for pod pending timeout (#857)
## Overview
We were previosuly using the `LastTrainsitionTime` for the phase transition on a pod pending timeout. This results in a stale "terminal" phase. In this PR, we use the current time instead.

## Test Plan
Tested locally.

## Rollout Plan (if applicable)
`managed-cluster-*-sync-all`

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [x] To be upstreamed to OSS

## Issue
https://linear.app/unionai/issue/BB-6201/phase-transition-startend-times-inconsistent-with-action-runtime

## Checklist
* [ ] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation