ctx/fix(propeller): ensure that request quotas are handled correctly (#866)
## Overview
It was discovered that the backoff handler was not considering GPU requests which lead to significant load on k8s when interacting with many GPU requests that could not be created due to quota violations.  It appears that there were 3 distinct issues at play here:

1. We were not matching anything more than simple alpha keys when parsing the quota errors.
2. We were only considering limits when making the ceiling calculations.
3. We did not collect requests when calculating the resource usage so we were missing GPU requests from the pod resource block.

This solves the issue by adding the correct matching to the regular expressions used to parse the quota errors.  When calculating the ceiling we now combine both the requests and the limits so that requests can now be used for consideration when checking for backoff conditions.  This also adds a change to gather pod resource requests.  Since the limits are guaranteed to be greater than the requests, we can gather the limits and then only add in requests when not specifically defined in the limits.

## Test Plan
Comprehensive tests were added.

## Rollout Plan (if applicable)
Deploy and test in staging, roll out to canary, and then update production.

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [x] To be upstreamed to OSS

## Issue
adhoc

## Checklist
* [ ] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation