ctx/fix(propeller): ignore missing inputs during abort (#868)
## Overview
Scaling issues exposed a problem where executions that never entered a started stated would fail abort.  This failure was due to the cleanup handler trying to retrieve the inputs.  Because jobs that are not yet started have run the initialization in the preExecute stage, the inputs do not exist and the abort continually fails in propeller.  It appears that these executions just get re-queued over and over again and are eventually cleaned up, but not until a good deal of time has passed - sometimes days. 

## Test Plan
* Test in staging and canary.

## Rollout Plan (if applicable)
* Merge
* Update submodule.

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [x] To be upstreamed to OSS

## Issue
adhoc

## Checklist
* [ ] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation