extend k8s plugins to support external gc (#867)
## Overview
Update k8s plugin interfaces to simplify garbage collection for v2. Note, the garage collector uses the functions in the GarbageCollectable field + the PluginEntry's ResourceToWatch.  Added a comment to leave as a hint if we ever run into issues in the future in which this isn't sufficient (ie a plugin creates resources that aren't managed by the resource to watch/deleted along w/ the resource to watch)

Will copy these over to the v2 plugins

## Test Plan
ran locally + in dogfood

## Rollout Plan (if applicable)
managed-cluster-all

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [] To be upstreamed to OSS

## Issue
ref: https://linear.app/unionai/issue/BB-6009/ensure-plugin-crs-are-cleaned-up

## Checklist
* [ ] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation