Add healthz path for vscode readiness probe (#877)
## Overview
Add a /healthz path for the VS Code readiness probe to bypass the standard heartbeat check, ensuring that the server does not remain active when the VS Code browser tab is closed.

## Test Plan
https://demo.hosted.unionai.cloud/v2/domain/development/project/flytesnacks/runs/rqx9chtwbg6dbcnxlbst?i=a0&tab=logs

## Rollout Plan (if applicable)
https://buildkite.com/unionai/managed-cluster-staging-sync/builds/10290#

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [ ] To be upstreamed to OSS

## Issue
https://unionai.slack.com/archives/C03SJ1VH8MV/p1765336129218729

## Checklist
* [ ] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation